Move to a React/Redux renderer
This commit introduces both React for rendering the Drupal Real-Time SEO
Widget as well as Redux for managing the state of the widget. This last
addition should make it easier to add multi-keyword support in the
future.

React was chosen because it makes it easier to control the markup than
chaining together strings and it also plays very well with Redux.

This new implementation doesn't expose the YoastSEO app anymore but
instead exposes a single Analyser that can be used for a form. In a
future commit this analyser will need to be slightly altered to be able
to handle multiple keyword and corresponding score storages.

The exposed Analyser class contains two methods that allow Drupal's form
controlling Javascript to signal the Redux store without exposing any
internals of the implementation, this should keep both sides of the
framework decoupled.