Complete the base for React Content Assessment
This fixes some issues in the translated readability assesments as well
as the ContentAssessor itself. This also hooks up the results of the
ContentAssessor to the Redux store. We do this in the most simple way
possible for now and can improve on this later.

Finally we build an AssessmentResult component to display each result
and wire up the Feedback component to the redux store to display the
content assessor feedback.