Add basic event handling to Analyser
This allows the changes in values from the redux store to be propagate
to Drupal using events. This ensures that the Analyser doesn't need to
know about Drupal's form layout (or whether a form is used at all).

This commit also modifies the SnippetPreview to actually update the
Redux store. The new EventEmitter that's part of the Analyser can be
used to propagate those changes from the redux store to Drupal.